+++ internal/jobworker/cgroup/cgroup_test.go
@@ -1,4 +1,4 @@
-package cgroups
+package cgroup
 
 import (
 	"errors"
