+++ internal/jobworker/cgroup/controller.go
@@ -1,4 +1,4 @@
-package cgroups
+package cgroup
 
 import (
 	"fmt"
